Rather than compute the checksum table statically, calculate the values on the fly. Additionally, add an lru cache for the results, with an upper bound on the cache size.

Updates the `test_print_table` test to use explicit output for the default crc_len=8, so that we can ensure that the above change doesn't affect the calculated values.

add higher value crc len options to `test_custom_positive_crc_len` - these would have caused memory errors previously.

resolves #90