@PowerBroker2 2024-04-17
Quick question: Is this going to impact real-time performance? I like how the lib currently precalculates the crc tables so that creating and checking the crc values only requires the processor to lookup values in a table quickly. Do these updates change that?