@robberwick 2024-04-18
This change is going to positively affect startup time, and also to a greater or lesser degree, the time taken to calculate the crc for a given payload of data.

It will positively affect the startup time by not pre-computing the checksum table for all possible values, which is a minor win for low values of crc length, but a much bigger win for value of crc length 32 and up (assuming your machine could even do that - attempting to instantiate a `CRC` object with a `crc_len` of 32 caused my machine to hang for the best part of a minute before python threw a `MemoryError`)

The calculation time is harder to quantify, as we compute the checksum for a given index on the fly, but then we also cache those results for the lifetime of the instance. So, there's a warming up period for the cache, but once the cache is fully primed, it's as though we're using a pre-computed lookup table. the only wrinkle on that is that the cache is implemented as an LRU cache with a max size of 65536 (2 ^ 32) items, so once we've got a cache with that many entries, it will start evicting the least recently used values when computing new ones. This is to mitigate against running into the same `MemoryError` issue when trying to use large values of `crc_len`.

In practice, it's probably not too much of an issue (certainly for values of crc_len <  32), once the cache is fully warmed up. For larger values. it seems like a decent tradeoff of runtime performance gainst being able to instantiate the objects without running out of machine memory.

before:

```
%timeit CRC()
174 µs ± 853 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
crc_8 = CRC()
%timeit crc_8.calculate([0x31, 0x32, 0x33, 0x34, 0x35])
712 ns ± 5.72 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
%timeit CRC(crc_len=16)
44.1 ms ± 227 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
crc_16 = CRC()
%timeit crc_16.calculate([0x31, 0x32, 0x33, 0x34, 0x35])
709 ns ± 1.6 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
%timeit CRC(crc_len=32)
<gave up after 30 minutes as it hadn't completed >
```

after:

```
%timeit CRC()
251 ns ± 28.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
crc_8 = CRC()
%timeit crc_8.calculate([0x31, 0x32, 0x33, 0x34, 0x35])
1.17 µs ± 12.8 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
%timeit CRC(crc_len=16)
303 ns ± 3.67 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
crc_16 = CRC()
%timeit crc_16.calculate([0x31, 0x32, 0x33, 0x34, 0x35])
1.16 µs ± 9.48 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
%timeit CRC(crc_len=32)
316 ns ± 2.78 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
crc_32 = CRC(crc_len=32)
%timeit crc_32.calculate([0x31, 0x32, 0x33, 0x34, 0x35])
1.14 µs ± 9.93 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```