@PowerBroker2 2024-04-18
Thanks! Looks like the live crc calc times aren't that bad - I'll merge